// frontend/src/generated/api/models/admin_adminTextureResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type admin_adminTextureResponse = {
    createdAt?: string;
    id?: number;
    layer?: 'woka' | 'body' | 'hair' | 'eyes' | 'hat' | 'accessory' | 'clothes' | 'companion';
    tags?: Array<string>;
    texture?: string;
    url?: string;
};